repaired flipped +% and %+ currency selector works again added file_type to media table (should work better now) fixed returning of an empty form, when error rises up removed delete button in country removed new button in review submitting review pointed on missing layout, fixed
git-svn-id: https://dev.virtuemart.net/svn/virtuemart/trunk/virtuemart@3326 87cc427b-4a1e-0410-b729-b2260ec0f063

+++ administrator/components/com_virtuemart/helpers/shopfunctions.php
@@ -27,14 +27,19 @@ class ShopFunctions {
 	}
 
 
-	function addStandardDefaultViewCommands (){
+	function addStandardDefaultViewCommands ($showNew=true,$showDelete=true){
 
 		JToolBarHelper::divider();
 		JToolBarHelper::publishList();
 		JToolBarHelper::unpublishList();
 		JToolBarHelper::editListX();
-		JToolBarHelper::addNewX();
-		JToolBarHelper::deleteList();
+		if($showNew){
+			JToolBarHelper::addNewX();
+		}
+		if($showDelete){
+			JToolBarHelper::deleteList();
+		}
+
 	}
 
 	function addStandardEditViewCommands (){
